+++ pythonFiles/include/blender_vscode/environment.py
@@ -1,11 +1,10 @@
+import os
+import sys
+from pathlib import Path
 from typing import Optional
 
-import bpy
-import sys
 import addon_utils
-from pathlib import Path
-import platform
-import os
+import bpy
 
 # binary_path_python was removed in blender 2.92
 # but it is the most reliable way of getting python path for older versions
@@ -19,3 +18,4 @@ scripts_folder = blender_path.parent / f"{version[0]}.{version[1]}" / "scripts"
 addon_directories = tuple(map(Path, addon_utils.paths()))
 
 EXTENSIONS_REPOSITORY: Optional[str] = os.environ.get("VSCODE_EXTENSIONS_REPOSITORY", "user_default") or "user_default"
+KEEP_ADDON_INSTALLED: bool = os.environ.get("VSCODE_KEEP_ADDON_INSTALLED") == "yes"