+++ pythonFiles/include/blender_vscode/operators/__init__.py
@@ -1,12 +1,9 @@
 from . import addon_update
 from . import script_runner
 from . import stop_blender
+from . import copy_string
 
-modules = (
-    addon_update,
-    script_runner,
-    stop_blender,
-)
+modules = (addon_update, script_runner, stop_blender, copy_string)
 
 
 def register():